CursorLoader and icons added
